refactor: 뷰 컨트롤러와, 데이터 반환 컨트롤러 분리
